feat(farm): implement farm CRUD and refactor project structure
- Added farmController with endpoints:
  • addFarm
  • getMyFarms
  • updateFarm
  • deleteFarm
- Updated Prisma schema:
  • Extracted farm fields from User model into new Farm model
  • Established one-to-many relation between User and Farm
- Refactored file naming convention for consistency:
  • Renamed exampleController.ts → example.controller.ts
  • Renamed exampleService.ts → example.service.ts
