+++ src/routes/prediction.routes.ts
@@ -1,0 +1,11 @@
+import { Router } from 'express';
+import { asyncWrapper } from '../middleware/asyncWrapper';
+import { createScan } from '../controllers/prediction.controller';
+import { uploadMemory } from '../middleware/upload';
+
+const router = Router();
+
+router.post('/create-scan', uploadMemory.single('image') ,  asyncWrapper(createScan));
+
+
+export default router;
